+++ Graphics/src/polygon/Hexlign.scala
@@ -2,6 +2,7 @@
 package ostrat
 package geom
 
+/** A regular Hexagon aligned to the X and Y axes. */
 trait Hexlign extends HexReg
 { def width: Double
   def height: Double
