+++ Util/src/Persist/PersistProduct.scala
@@ -2,8 +2,8 @@
 package ostrat
 import pParse._
 
-/** The base trait for the persistence of Case classes, aka Product types */
-trait PersistCase[R] extends ShowCaseT[R] with PersistCompound[R]
+/** The base trait for the persistence of algebraic product types, including case classes. */
+trait PersistProduct[R] extends ShowProductT[R] with PersistCompound[R]
 {
   override def fromExpr(expr: ParseExpr): EMon[R] =  expr match
   {
@@ -13,10 +13,10 @@ trait PersistCase[R] extends ShowCaseT[R] with PersistCompound[R]
   }
 }
 
-/** Persistence class for 2 parameter case classes. */ 
+/** Persistence class for 2 parameter case classes. */
 class Persist2[A1, A2, R](typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, val newT: (A1, A2) => R,
   opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit ev1: Persist[A1], ev2: Persist[A2], eq1: Eq[A1], eq2: Eq[A2]) extends
-  Show2T[A1, A2, R](typeStr, name1, fArg1, name2, fArg2, opt2, opt1) with PersistCase[R]
+  Show2T[A1, A2, R](typeStr, name1, fArg1, name2, fArg2, opt2, opt1) with PersistProduct[R]
 {
   //override def fromClauses(clauses: Refs[Clause]): EMon[R] = fromClauses2(newT, clauses)
  /* override def fromParameterStatements(sts: Refs[Statement]): EMon[R] = (sts, opt1, opt2) match
@@ -42,11 +42,11 @@ class PersistInt2[R](typeStr: String, name1: String, fArg1: R => Int, name2: Str
 class PersistD2[R](typeStr: String, name1: String, fArg1: R => Double, name2: String, fArg2: R => Double, newT: (Double, Double) => R) extends
    Persist2[Double, Double, R](typeStr, name1, fArg1, name2, fArg2, newT)
 
-/** Persistence class for 3 parameter case classes. */   
+/** Persistence class for 3 parameter case classes. */
 class Persist3[A1, A2, A3, R](typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String, fArg3: R => A3,
   val newT: (A1, A2, A3) => R, opt3: Option[A3] = None, opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit ev1: Persist[A1], ev2: Persist[A2],
   ev3: Persist[A3], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3]) extends
-  Show3T[A1, A2, A3, R](typeStr, name1, fArg1, name2, fArg2, name3, fArg3, opt3, opt2, opt1) with PersistCase[R]
+  Show3T[A1, A2, A3, R](typeStr, name1, fArg1, name2, fArg2, name3, fArg3, opt3, opt2, opt1) with PersistProduct[R]
 {
   //override def fromClauses(clauses: Refs[Clause]): EMon[R] = fromClauses3(newT, clauses)
   /*override def fromParameterStatements(sts: Refs[Statement]): EMon[R] = (sts, opt1, opt2, opt3) match
@@ -71,11 +71,11 @@ object Persist3
 abstract class PersistD3[R](typeStr: String, name1: String, fArg1: R => Double, name2: String, fArg2: R => Double, name3: String, fArg3: R => Double,
   newT: (Double, Double, Double) => R) extends Persist3[Double, Double, Double, R](typeStr, name1, fArg1, name2, fArg2, name3, fArg3, newT)
 
-/** Persistence class for 4 parameter case classes. */   
+/** Persistence class for 4 parameter case classes. */
 class Persist4[A1, A2, A3, A4, R](typeStr: String, name1: String, fArg1: R => A1, name2: String, fArg2: R => A2, name3: String, fArg3: R => A3,
   name4: String, fArg4: R => A4, val newT: (A1, A2, A3, A4) => R, opt4: Option[A4], opt3: Option[A3] = None, opt2: Option[A2] = None,
   opt1: Option[A1] = None)(implicit ev1: Persist[A1], ev2: Persist[A2], ev3: Persist[A3], ev4: Persist[A4], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3],
-  eq4: Eq[A4]) extends Show4T(typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, opt4, opt3, opt2, opt1) with PersistCase[R]
+  eq4: Eq[A4]) extends Show4T(typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, opt4, opt3, opt2, opt1) with PersistProduct[R]
 {
  // override def fromClauses(clauses: Refs[Clause]): EMon[R] = fromClauses4(newT, clauses)
   /*override def fromParameterStatements(sts: Refs[Statement]): EMon[R] = (sts, opt1, opt2, opt3, opt4) match
@@ -106,7 +106,7 @@ class Persist5[A1, A2, A3, A4, A5, R](typeStr: String, name1: String, fArg1: R =
   name4: String, fArg4: R => A4, name5: String, fArg5: R => A5, val newT: (A1, A2, A3, A4, A5) => R, opt5: Option[A5], opt4: Option[A4] = None,
   opt3: Option[A3] = None, opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit ev1: Persist[A1], ev2: Persist[A2], ev3: Persist[A3],
   ev4: Persist[A4], ev5: Persist[A5], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5]) extends
-  Show5T(typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, name5, fArg5, opt5, opt4, opt3, opt2, opt1) with PersistCase[R]
+  Show5T(typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, name5, fArg5, opt5, opt4, opt3, opt2, opt1) with PersistProduct[R]
 {
  // override def fromClauses(clauses: Refs[Clause]): EMon[R] = fromClauses5(newT, clauses)
   /*override def fromParameterStatements(sts: Refs[Statement]): EMon[R] = (sts, opt1, opt2, opt3, opt4, opt5) match
@@ -148,7 +148,7 @@ class Persist6[A1, A2, A3, A4, A5, A6, R](typeStr: String, name1: String, fArg1:
   opt2: Option[A2] = None, opt1: Option[A1] = None)(implicit ev1: Persist[A1], ev2: Persist[A2], ev3: Persist[A3], ev4: Persist[A4], ev5: Persist[A5],
   ev6: Persist[A6], eq1: Eq[A1], eq2: Eq[A2], eq3: Eq[A3], eq4: Eq[A4], eq5: Eq[A5], eq6: Eq[A6]) extends
   Show6T(typeStr, name1, fArg1, name2, fArg2, name3, fArg3, name4, fArg4, name5, fArg5, name6, fArg6, opt6, opt5, opt4, opt3, opt2, opt1) with
-  PersistCase[R]
+  PersistProduct[R]
 {
   //override def fromClauses(clauses: Refs[Clause]): EMon[R] = fromClauses6(newT, clauses)
   /*override def fromParameterStatements(sts: Refs[Statement]): EMon[R] = (sts, opt1, opt2, opt3, opt4, opt5, opt6) match
